+++ TeamCode/src/main/java/org/betastudio/ftc/thread/InfinityLoopThread.java
@@ -1,7 +1,6 @@
 package org.betastudio.ftc.thread;
 
 import org.betastudio.ftc.Interfaces;
-import org.betastudio.ftc.action.Action;
 import org.betastudio.ftc.action.Actions;
 
 public class InfinityLoopThread extends Thread implements Interfaces.ThreadEx {
@@ -21,23 +20,4 @@ public class InfinityLoopThread extends Thread implements Interfaces.ThreadEx {
 		action.closeTask();
 	}
 
-	public static final class InfinityLoopAction implements Action, Interfaces.ThreadEx {
-		private final Runnable runnable;
-		private       boolean  interrupted;
-
-		public InfinityLoopAction(final Runnable runnable) {
-			this.runnable = runnable;
-		}
-
-		@Override
-		public boolean activate() {
-			runnable.run();
-			return ! interrupted;
-		}
-
-		@Override
-		public void closeTask() {
-			interrupted = true;
-		}
-	}
 }
