+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/eventloop/integral/IntegralTeleOp.java
@@ -45,8 +45,8 @@ public abstract class IntegralTeleOp extends OverclockOpMode implements Integral
 		client.setUpdateConfig(UpdateConfig.MANUALLY);
 
 		final MethodFrequencyCaller caller = new MethodFrequencyCaller(client::update);
-		caller.setRequestCaller(() -> is_terminate_method_called || isStopRequested());
-		caller.setFrequencyFPS(10);
+		caller.setStopRequestCaller(() -> is_terminate_method_called || isStopRequested());
+		caller.setFrequencyFPS(5);
 		Global.service.execute(caller);
 
 		HardwareDatabase.sync(hardwareMap, true);
